# PERF: faster unstacking